backend/requirements.txt
Flask==3.0.2
flask-sock==0.7.0
simple-websocket==1.0.0
gunicorn==21.2.0
PyYAML==6.0.1
